Refactoring colab.js and Nebari.js for common code in creating notebook cells
