+++ functions/utils.js
@@ -103,3 +103,73 @@ export function getRootUrlWithoutTrailingSlash(url) {
 
   return rootUrl
 }
+
+/**
+ * Get the correct formatted notebook for Colab and Nebari functions
+ * @param {string} title
+ * @param {string} zipRes
+ * @param {string} argparser
+ * @param {string} template
+ * @returns {JSON} nb
+ */
+export function getNbCells(title, zipRes, argparser, template) {
+  function create_nb_cell(source_array, cell_type) {
+    return {
+      cell_type: cell_type,
+      metadata: {},
+      execution_count: null,
+      outputs: [],
+      source: source_array
+    }
+  }
+
+  let specific_commands = []
+
+  if (title === 'Template Vision Segmentation') {
+    specific_commands.push(
+      '!python -c "from data import download_datasets; download_datasets(\'./\')"'
+    )
+  }
+
+  const md_cell = [
+    `# ${title} by PyTorch-Ignite Code-Generator\n\n`,
+    'Please, run the cell below to execute your code.'
+  ]
+
+  const common_nb_commands = [
+    `!wget ${zipRes}\n`,
+    `!unzip ${template}.zip\n`,
+    '!pip install -r requirements.txt'
+  ]
+
+  const execution_nb_commands = [
+    `!python main.py ${
+      argparser === 'hydra'
+        ? '#--config-dir=/content/ --config-name=config.yaml'
+        : 'config.yaml'
+    }`
+  ]
+
+  let nb_cells = [
+    create_nb_cell(md_cell, 'markdown'),
+    create_nb_cell(common_nb_commands, 'code')
+  ]
+  if (specific_commands.length > 0) {
+    nb_cells.push(create_nb_cell(specific_commands, 'code'))
+  }
+  nb_cells.push(create_nb_cell(execution_nb_commands, 'code'))
+
+  const nb = {
+    nbformat: 4,
+    nbformat_minor: 0,
+    metadata: {
+      kernelspec: {
+        display_name: 'Python 3',
+        name: 'python3'
+      },
+      accelerator: 'GPU'
+    },
+    cells: nb_cells
+  }
+  return nb
+}
